Moved mapMember out of mapper file
no-issue

This cleans up the serializer a bit by keeping it's functionality all in
one place, rather than a shared mapper file
