+++ core/server/api/canary/utils/serializers/output/utils/mapper.js
@@ -142,23 +142,6 @@ const mapAction = (model, frame) => {
     return attrs;
 };
 
-const mapMember = (model, frame) => {
-    const jsonModel = model.toJSON ? model.toJSON(frame.options) : model;
-
-    if (_.get(jsonModel, 'stripe.subscriptions')) {
-        let compedSubscriptions = _.get(jsonModel, 'stripe.subscriptions').filter(sub => (sub.plan.nickname === 'Complimentary'));
-        const hasCompedSubscription = !!(compedSubscriptions.length);
-
-        // NOTE: `frame.options.fields` has to be taken into account in the same way as for `stripe.subscriptions`
-        //       at the moment of implementation fields were not fully supported by members endpoints
-        Object.assign(jsonModel, {
-            comped: hasCompedSubscription
-        });
-    }
-
-    return jsonModel;
-};
-
 const mapLabel = (model, frame) => {
     const jsonModel = model.toJSON ? model.toJSON(frame.options) : model;
     return jsonModel;
@@ -189,5 +172,4 @@ module.exports.mapIntegration = mapIntegration;
 module.exports.mapSettings = mapSettings;
 module.exports.mapImage = mapImage;
 module.exports.mapAction = mapAction;
-module.exports.mapMember = mapMember;
 module.exports.mapEmail = mapEmail;
